Remove unused code, move LatestSnapshotsMetadataLoader
